e2e/testdata/marketing-strategy/pyproject.toml
[tool.poetry]
name = "marketing-strategy-multi"
version = "0.1.0"
description = ""
authors = ["Mauro Sardara <msardara@cisco.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
langchain-ollama = "^0.1.3"
crewai-tools = "^0.12.0"
traceloop-sdk = "^0.30.0"
crewai = "^0.55.2"
arize-otel = "^0.5.1"
openinference-instrumentation-crewai = "^0.1.2"
openinference-instrumentation-langchain = "^0.1.28"
duckduckgo-search = "^6.2.12"
langgraph = "^0.2.28"
kubernetes = "^31.0.0"
dash = "^2.18.1"
flask = "^3.0.3"
dash-cytoscape = "^1.0.2"
jinja2 = "^3.1.4"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"

[tool.poetry.scripts]
marketing_strategy = "marketing_strategy_multi.main:run"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
